+++ frontend/src/components/userpage/TabBar.module.css
@@ -1,5 +1,18 @@
 .tabBar{
+    width: 90%;
+    margin: 20px 5px 5px 5px;
+    padding: 1vw 2vw;
+    border: 2px solid var(--highlight-color);
+    border-radius: 10px;
     display: flex;
-    flex-wrap: wrap;
-    border: 1px solid #e0e0e0;
+    align-items: center;
+    justify-content: space-between;
+    box-sizing: border-box;
 }
+
+.tabBar button{
+    margin: 0px 5px;
+    padding: 10px;
+    width: 15%;
+}
+
